Fix the fix command (#23)

+++ src/note_clerk/fixing.py
@@ -6,13 +6,17 @@ import re
 from typing import Any, Callable, Dict, Iterable, Optional, TextIO, Tuple, Union
 
 from boltons.fileutils import atomic_save
+import click
 from dateutil.parser import parse as parse_date
 from orderedset import OrderedSet
 from ruamel.yaml import YAML
+from ruamel.yaml.constructor import DuplicateKeyError
+from ruamel.yaml.parser import ParserError
+from ruamel.yaml.scanner import ScannerError
 from ruamel.yaml.timestamp import TimeStamp
 
 from . import utils
-from .utils import ensure_newline
+from .utils import ensure_newline, UnclosedHeader
 
 
 log = logging.getLogger(__name__)
@@ -48,9 +52,10 @@ def merge_values(key: str, existing: Any, new: Any) -> Any:
     if key == "created":
         try:
             log.debug(f"{type(existing)=} {type(new)=}")
+            log.debug(f"{existing=} {new=}")
             return min_date(existing, new)
-        except (TypeError, ValueError):
-            pass
+        except TypeError:
+            raise UnableFix("Can't identify timezones")
     if isinstance(existing, int) or isinstance(new, int):
         raise UnableFix("Unable to join integers")
     raise UnableFix("Unable to join constants")
@@ -60,14 +65,14 @@ def fix_header(header: str) -> str:
     output = io.StringIO()
     yaml = YAML(output=output)
     header_docs = [h for h in yaml.load_all(header) if h is not None]  # noqa: S506
-    log.debug(f"header_docs:\n{header_docs}")
 
     if len(header_docs) < 2:
-        return header
+        if len(header) == 0:
+            return header
+        return ensure_newline(header)
 
     combined: Dict[str, Any] = {}
     for doc in header_docs:
-        log.debug(f"doc:\n{doc}")
         for key, value in sorted(doc.items(), key=lambda t: t[0]):
             try:
                 combined[key] = merge_values(key, combined[key], value)
@@ -75,8 +80,7 @@ def fix_header(header: str) -> str:
                 combined[key] = value
 
     yaml.dump(combined, output)
-
-    return f"---\n{output.getvalue().strip()}\n***\n"
+    return f"---\n{output.getvalue().strip()}\n---\n"
 
 
 ID_REGEX = re.compile(r"^([0-9]+)")
@@ -88,7 +92,6 @@ def fix_filename(filename: Optional[str]) -> Optional[str]:
     path = Path(filename)
     stem = path.stem
     match = ID_REGEX.match(stem)
-    log.debug(f"{match=}")
     if match is None:
         return filename
     orig_note_id = match.groups()[0]
@@ -112,24 +115,36 @@ def fix_filename(filename: Optional[str]) -> Optional[str]:
 
 
 def fix_text(text: TextIO, filename: Optional[str]) -> Tuple[str, Optional[str]]:
-    header, body = utils.split_header([line.strip() for line in text.readlines()])
     try:
+        header, body = utils.split_header(
+            [line.removesuffix("\n") for line in text.readlines()]
+        )
         new_header = fix_header(header)
-    except Exception:
-        log.error("error creating header", exc_info=True)
+    except UnableFix:
         raise
-    log.debug(f"new_header:\n{new_header}")
-    new_note: str = ensure_newline(new_header) + ensure_newline(body.strip())
+    except (ParserError, ScannerError) as e:
+        raise UnableFix("Malformed header") from e
+    except DuplicateKeyError as e:
+        raise UnableFix("Duplicate Key found in header document") from e
+    except UnclosedHeader as e:
+        raise UnableFix("Unclosed header") from e
+    except UnicodeDecodeError as e:  # pragma: no cover
+        raise UnableFix("Invalid File") from e
+    except Exception as e:  # pragma: no cover
+        log.error(f"error creating header for {filename}", exc_info=True)
+        raise UnableFix("Unknown Error") from e
+    new_note: str = new_header + ensure_newline(body)
     new_filename = fix_filename(filename)
     return new_note, new_filename
 
 
 def raised_error(func: Callable) -> Callable:
-    def wrapper(*args: Any, **kwargs: Any) -> Iterable[bool]:
+    def wrapper(text: TextIO, filename: Optional[str]) -> Iterable[bool]:
         try:
-            func(*args, **kwargs)
+            func(text, filename)
             yield False
-        except UnableFix:
+        except UnableFix as e:
+            log.warning(f"Unable to fix '{filename or 'stdin'}': {e} ")
             yield True
 
     return wrapper
@@ -138,13 +153,16 @@ def raised_error(func: Callable) -> Callable:
 @raised_error
 def update_text(
     text: TextIO,
-    filename: str,
+    filename: Optional[str],
 ) -> None:
     log.debug(f"{filename=}")
     n_text, n_filename = fix_text(text, filename)
-    with atomic_save(n_filename, overwrite=True) as f:
-        f.write(n_text.encode("utf-8"))
-    log.debug(f"\n  {filename=}\n{n_filename=}")
-    if filename is not None and filename != n_filename:
-        log.debug("Deleting file")
-        Path(filename).unlink()
+
+    if n_filename is None:
+        click.echo(n_text, nl=False)
+    else:
+        with atomic_save(n_filename, overwrite=True) as f:
+            f.write(n_text.encode("utf-8"))
+        if filename is not None and filename != n_filename:
+            log.debug(f"Deleting file: {filename}")
+            Path(filename).unlink()
